% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_titled_wb function using openxlsx.R
\name{create_titled_wb}
\alias{create_titled_wb}
\title{create_titled_wb - Create a Workbook with Multiple Titled Worksheets Using 'openxlsx'}
\usage{
create_titled_wb(filename = "Test", dataframe_list,
  titles = LETTERS[1:length(dataframe_list)],
  tabnames = LETTERS[1:length(dataframe_list)],
  format_column_labels = TRUE, ws_type = "table",
  output_path = getwd(), open_wb = TRUE, overwrite_existing = TRUE)
}
\arguments{
\item{filename}{The worksheet name you would like, including .xlsx file extension.}

\item{dataframe_list}{A list of dataframes (without quotes) created outside the function call.}

\item{titles}{A vector of worksheet titles, 1 for each worksheet tab.}

\item{tabnames}{A vector of tabnames, current maximum # tabs supported = 8.}

\item{format_column_labels}{TRUE or FALSE.}

\item{ws_type}{"table" or "standard".}

\item{output_path}{Location where you want the workbook to be saved.}

\item{open_wb}{If TRUE, open the workbook for review.}
}
\description{
create_titled_wb - Create a Workbook with Multiple Titled Worksheets Using 'openxlsx'
}
\examples{
library(WorksheetFunctions)
df1 <- smutilities::create_tbl_df(fill_value=0, nrows=10, ncols=10)
df2 <- smutilities::create_tbl_df(fill_value="A2", nrows=10, ncols=10)
df3 <- smutilities::create_tbl_df(fill_value="Test", nrows=10, ncols=10)
stub_colnames <- rep(c("Test.Name", "Test_Name", "test.Name",
                       "test_NAMe", "test.NAME"), 2)
stub_colnames <- paste(stub_colnames, 1:length(stub_colnames), sep=" ")
names(df1) <- stub_colnames
names(df2) <- stub_colnames
names(df3) <- stub_colnames

dataframes <- list(df1, df2, df3)

create_titled_wb(filename = "Test_New",
                 dataframe_list = dataframes,
                 titles = paste("Title ", 1:length(dataframe_list)),
                 tabnames = LETTERS[1:length(dataframe_list)],
                 format_column_labels = TRUE,
                 ws_type = "table",
                 output_path = getwd(),
                 open_wb = TRUE,
                 overwrite_existing = TRUE)
}
\keyword{xlsx}
