% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_worksheetfunctions.R
\name{update_worksheetfunctions}
\alias{update_worksheetfunctions}
\title{update_worksheetfunctions - Update, Compile, and Re-Install the 'WorksheetFunctions' package.}
\usage{
update_worksheetfunctions()
}
\description{
update_worksheetfunctions - Update, Compile, and Re-Install the 'WorksheetFunctions' package.
}
\examples{
library(WorksheetFunctions)
update_worksheetfunctions()
}
\keyword{WorksheetFunctions,}
\keyword{xlsx}

